Use interactice service to start openvpn process
Instead of starting openvpn process directly:

 - connect to interactive service pipe
 - send startup info
 - read openvpn process pid (for polling)

Since we use virtual service account, we don't have
privileges to subscribe for process events, so instead
we use async polling to check if process is alive and
restart it of it is not.

Removed Stop() since due to lack of privileges
the only way now we can stop the process is to
signal the event.

GitHub: #23

Signed-off-by: Lev Stipakov <lev@openvpn.net>